#41 - Fix a type error from my over-zealous search/replace.

+++ chapter-36-layers/src/map_builders/simple_map.rs
@@ -39,7 +39,7 @@ impl SimpleMapBuilder {
 
                 if !rooms.is_empty() {
                     let (new_x, new_y) = new_room.center();
-                    let (prev_x, prev_y) = rooms[i-1].center();
+                    let (prev_x, prev_y) = rooms[i as usize -1].center();
                     if rng.range(0,2) == 1 {
                         apply_horizontal_tunnel(&mut build_data.map, prev_x, new_x, prev_y);
                         apply_vertical_tunnel(&mut build_data.map, prev_y, new_y, new_x);
